correction to dimension_reduction tests
Corrected test/dimension_reduction.jl so that the tests should all pass now.

After further review, I concluded that the mean should not be subtracted from the variables d in DimensionReduction/pca.jl in our instance, where we wish to find a subspace W such that (W')*d gives the reduced dimensional form y, and W*y = d (recovers the original variable) (see the function fit_transform()). In other words, we would like for the first eigenvector of W to retain the mean trend of d, rather than lose that info. This results in a form of PCA where the first principal component does not correspond to max variance, but rather the second one.